+++ src/Server/Helper.php
@@ -610,13 +610,10 @@ class Helper
      */
     private function doConvertToPsrResponse($result, ResponseInterface $response, StreamInterface $writableBodyStream): ResponseInterface
     {
-        $httpStatus = $this->resolveHttpStatus($result);
-
-        $result = json_encode($result, JSON_THROW_ON_ERROR);
-        $writableBodyStream->write($result);
+        $writableBodyStream->write(json_encode($result, JSON_THROW_ON_ERROR));
 
         return $response
-            ->withStatus($httpStatus)
+            ->withStatus($this->resolveHttpStatus($result))
             ->withHeader('Content-Type', 'application/json')
             ->withBody($writableBodyStream);
     }